requirements.txt
python-dotenv
GitPython
PyGithub